refactor: Enable transfer transaction retrieval from appwrite
This commit enables the retrieval of transfer transactions from appwrite by uncommenting the necessary code. The transfer transactions are obtained using the getTransactionsByBankId function from the transaction.actions module. This change allows for the inclusion of transfer transactions in the list of all transactions, which are then sorted by date.